# String escape & unescape unit tests & optimizations