@jwilder 2016-11-29
Can you run the before and after for the `models` package?  The `Unescape` allocation might be more significant using those benchmarks.