@sparrc 2016-12-05
Sorry let this fall through the cracks. Doesn't look like any of the tests in the models package does any unescaping:

```
master:

$ go test ./models/... --bench=. --benchmem
BenchmarkMarshal-8                     	 2000000	       924 ns/op	     416 B/op	       4 allocs/op
BenchmarkNewPoint-8                    	  300000	      4566 ns/op	    3424 B/op	      28 allocs/op
BenchmarkParsePointNoTags5000-8        	    1000	   1666368 ns/op	  72.01 MB/s	 1644800 B/op	    5002 allocs/op
BenchmarkParsePointNoTags-8            	 3000000	       397 ns/op	  57.85 MB/s	     336 B/op	       3 allocs/op
BenchmarkParsePointWithPrecisionN-8    	 5000000	       385 ns/op	  59.67 MB/s	     336 B/op	       3 allocs/op
BenchmarkParsePointWithPrecisionU-8    	 5000000	       390 ns/op	  58.88 MB/s	     336 B/op	       3 allocs/op
BenchmarkParsePointsTagsSorted2-8      	 3000000	       530 ns/op	  96.09 MB/s	     368 B/op	       3 allocs/op
BenchmarkParsePointsTagsSorted5-8      	 2000000	       700 ns/op	 118.47 MB/s	     400 B/op	       3 allocs/op
BenchmarkParsePointsTagsSorted10-8     	 1000000	      1004 ns/op	 142.40 MB/s	     448 B/op	       3 allocs/op
BenchmarkParsePointsTagsUnSorted2-8    	 2000000	       666 ns/op	  76.48 MB/s	     400 B/op	       4 allocs/op
BenchmarkParsePointsTagsUnSorted5-8    	 2000000	       964 ns/op	  86.01 MB/s	     464 B/op	       4 allocs/op
BenchmarkParsePointsTagsUnSorted10-8   	 1000000	      1675 ns/op	  85.37 MB/s	     576 B/op	       4 allocs/op
BenchmarkParseKey-8                    	 3000000	       491 ns/op	     611 B/op	       3 allocs/op
PASS
ok  	github.com/influxdata/influxdb/models	26.594s

PR:

$ go test ./models/... --bench=. --benchmem
BenchmarkMarshal-8                     	 2000000	       910 ns/op	     416 B/op	       4 allocs/op
BenchmarkNewPoint-8                    	  300000	      3928 ns/op	    3424 B/op	      28 allocs/op
BenchmarkParsePointNoTags5000-8        	    1000	   1663142 ns/op	  72.15 MB/s	 1644800 B/op	    5002 allocs/op
BenchmarkParsePointNoTags-8            	 3000000	       399 ns/op	  57.62 MB/s	     336 B/op	       3 allocs/op
BenchmarkParsePointWithPrecisionN-8    	 5000000	       386 ns/op	  59.47 MB/s	     336 B/op	       3 allocs/op
BenchmarkParsePointWithPrecisionU-8    	 5000000	       393 ns/op	  58.51 MB/s	     336 B/op	       3 allocs/op
BenchmarkParsePointsTagsSorted2-8      	 3000000	       532 ns/op	  95.69 MB/s	     368 B/op	       3 allocs/op
BenchmarkParsePointsTagsSorted5-8      	 2000000	       700 ns/op	 118.50 MB/s	     400 B/op	       3 allocs/op
BenchmarkParsePointsTagsSorted10-8     	 1000000	      1002 ns/op	 142.59 MB/s	     448 B/op	       3 allocs/op
BenchmarkParsePointsTagsUnSorted2-8    	 2000000	       667 ns/op	  76.40 MB/s	     400 B/op	       4 allocs/op
BenchmarkParsePointsTagsUnSorted5-8    	 2000000	       965 ns/op	  85.94 MB/s	     464 B/op	       4 allocs/op
BenchmarkParsePointsTagsUnSorted10-8   	 1000000	      1704 ns/op	  83.91 MB/s	     576 B/op	       4 allocs/op
BenchmarkParseKey-8                    	 3000000	       490 ns/op	     611 B/op	       3 allocs/op
PASS
ok  	github.com/influxdata/influxdb/models	26.418s
```